refactor: Update directory structure with components and utils
