+++ src/components/Panel.js
