+++ src/components/Pins.js
