+++ src/utils/constants/cities.json
